Rickettsiella endosymbiont information

An endosym object defining the life history for (Rickettsiella). Can be used to run the endosymbiont model on GPA.

fun_fitness fitted using fit_null to be fixed at 0.5 based on Gu et al. 2023
fun_trans_eff fitted using fit_quadratic to peak around 20°C based Gu et al. 2023
fun_susc fitted using fit_quadratic to peak around 20°C based Gu et al. 2023
fun_vert_trans fitted using fit_null to be fixed at 1 based on Gu et al. 2023
introduction_date is 2022-04-17, a week after crop emergence
introduction_n is 300 individuals

Usage

data(Rickettsiella)

Format

An object of class endosym; see endosym.

References

Gu, X., Ross, P.A., Gill, A., Yang, Q., Ansermin, E., Sharma, S., Soleimannejad, S., Sharma, K., Callahan, A., Brown, C., Umina, P.A., Kristensen, T.N., Hoffmann, A.A. A rapidly spreading deleterious aphid endosymbiont that uses horizontal as well as vertical transmission. PNAS, 12, e2217278120 (2023)