Counterfactual Simulations

Run multiple simulations of the endosymbiont model under different scenarios

Usage

counterfact(
  Pest,
  Endosymbiont,
  Crop,
  Parasitoid = NULL,
  init,
  conds,
  modules = NULL
)

Arguments

Pest

object of class pest defining the pest

Endosymbiont

object of class endosym defining the endosymbiont

Crop

object of class crop defining the crop

Parasitoid

object of class parasitoid defining the parasitoid

init

object of class initial defining the starting conditions for the simulation

conds

object of class sim_conds defining the simulation conditions

modules

list of modules and scenarios to run, if NULL (default) all fixed to FALSE

Value

object of class endosim_col; see endosim_col.

Details

Uses the endosim function to run simulations of the endosymbiont model. A basic null model is always run, without any modules. The provided pest, endosym, crop,parasitoid, initial, and sim_conds objects are used to parameterise the simulation.

modules should be a list with the following arguments: vert_trans, hori_trans, imi, emi, para. Each argument should be a logical vector, consisting either of FALSE (fixed off), TRUE (fixed on), or c(FALSE, TRUE) (run counterfactual scenarios with module turned on and turned off).

See also