EndoSim

Run simulation using the endosymbiont model

Usage

endosim(
  Pest,
  Endosymbiont,
  Crop,
  Parasitoid = NULL,
  init,
  conds,
  plot = TRUE,
  progress = TRUE,
  vert_trans = FALSE,
  hori_trans = FALSE,
  imi = FALSE,
  emi = FALSE,
  para = FALSE
)

Arguments

Pest

object of class pest defining the pest

Endosymbiont

object of class endosym defining the endosymbiont

Crop

object of class crop defining the crop

Parasitoid

object of class parasitoid defining the parasitoid

init

object of class initial defining the starting conditions for the simulation

conds

object of class sim_conds defining the simulation conditions

plot

if TRUE (default) generate plots

progress

if TRUE (default) print progress bar

vert_trans

set to TRUE to turn on vertical transmission

hori_trans

set to TRUE to turn on horizontal transmission

imi

set to TRUE to turn on immigration into paddock

emi

set to TRUE to turn on emigration from paddock

para

set to TRUE to turn on parasitoids

Value

object of class endosim_mod; see endosim_mod.

Details

Constructs the endosymbiont model using the provided pest, endosym, crop,parasitoid, initial, and sim_conds objects.

The Pest species feeds on the provided Crop.

Vertical and horizontal transmission of endosymbiont are defined using functions in the Endosymbiont.

Parasitism rates are defined usined functions in the Parasitoid.