An S4 class to store a pest's life history for the endosymbiont model

An S4 class to store a pest's life history for the endosymbiont model

Slots

species

The name of the pest

bg_loss

Background mortality (0-1)

alate_penalty

Relative fecundity penalty of alates compared to apterae (0-1)

apterae_walk

Number of plants per day visited by apterae

alate_flight

Number of plants per day visited by alates

fun_dev_apt

Function describing temperature-dependent development of nymphs destined to become apterae

fun_dev_ala

Function describing temperature-dependent development of nymphs destined to become alates

fun_imi_neg

Function describing number of immigrant R- alates

fun_imi_pos

Function describing number of immigrant R+ alates

fun_emi

Function describing number of emigrants among newly metamorphosed alates

fun_temp_loss

Function describing daily mortality due to exceeding temperature thresholds

fun_rainfall_loss

Function describing daily mortality due to rainfall

fun_sen_loss

Function describing daily mortality due to senescence

fun_dens_fecund

Function describing daily fecundity as function of pest density

fun_temp_fecund

Function describing daily fecundity as function of temperature

fun_age_fecund

Function describing daily fecundity as function of adult cohort age

fun_alate_prod

Function describing proportion of newly produced nymphs destined to become alates as function of pest density

fun_damage

Function describing reduction in crop biomass due to pest density

Examples

GPA <-
new("pest",
  species = "Myzus persicae",
  bg_loss = 0.03,
  alate_penalty = 0.5,
  apterae_walk = 0.0005,
  alate_flight = 0.005,
  fun_dev_apt = EndoSim:::fit_wang(24.1, 0.152, 18.09, 32.79, 2.84, 1.8977),
  fun_dev_ala = EndoSim:::fit_wang(24.1, 0.16474, 19.96, 33.17, 1.41, 2.2202),
  fun_imi_neg = EndoSim:::fit_imi_barton(),
  fun_imi_pos = EndoSim:::fit_null(0),
  fun_emi = EndoSim:::fit_emi_thack(),
  fun_temp_loss = EndoSim:::fit_supergaus(32.79, 2.84, 0.00083, 0.00017, 7.66, 17.66),
  fun_rainfall_loss = EndoSim:::fit_rainfall_thack(),
  fun_sen_loss = EndoSim:::fit_gompertz(0.00188, 0.13),
  fun_dens_fecund = EndoSim:::fit_bannerman(10000, 0.0008),
  fun_temp_fecund = EndoSim:::fit_quadratic(-0.7611, 31.9847, 25),
  fun_age_fecund = EndoSim:::fit_weibull(58, 1.885, 5.953, 0),
  fun_alate_prod = EndoSim:::fit_alate(67.418, 0.993, 0.076, 300),
  fun_damage = EndoSim:::fit_null(0)
 )

GPA
#> Pest of the species  Myzus persicae